Improved loadout control system
Loadouts can be defined in a much more user-friendly manner with categories that only have to be defined once. Script automatically goes through the categories assigned to an item's slots (through the whole depth) and adds them to the mods record in the bot db. Common json files can be made that will be used for all loadout processing, reducing repeat definitions.